Make spectacle a lua module.
In order to function as a module, need to put exported functions into a
table and return that table at the end of the module. However pandoc
does not load custom writers as a module but simply loads the script
code, so need to account for that by checking for a nil module name and
copying the exported functions into the global namespace.